feat (Adicionando mudanças de layout e função de excluir monitoria - Adicionado ícones de 'erro' associado com as mensagens de 'not found', icones salvos na pasta assets como monitoramento e monitoramento2. - Adicionando nova página de dashboard de admin que contém os cards que levam para as páginas das entidades que o admin pode alterar - Adicionado novo ícone de 'setinha' ao lado da aba 'Minhas monitorias' - Adicinada animação na página de perfil que, ao alterar qualquer valor dos inputs das informações, uma borda azul indica que o campo foi alterado - Adicionado ícone de lixeira na aba 'gerenciamento' para que seja possível excluir uma monitoria - Adicionado lógica de implementção para a função de delete de monitoria, permitindo somente o autor da monitoria ou o admin - Adicionando interface UsuarioLogado para retornos nas funções que usam a interface 'AuthRequest 'extendida do Request definido no middleware de autenticação - Adicionado evento 'Monitoria:Excluir', listener de evento e registro na tabela de audiotoria - Adicionado evento 'Aluno:Senha', listener de evento e registro na tabela de audiotoria que indica alteração de senha - Adicionado fix no frontend da função de atulizar senha que não pegava o outro campo 'senhaConfirmada', e impedia a atualização da senha via frontend)

+++ src/services/Monitoria/MonitoriaService.ts
@@ -2,6 +2,7 @@ import { json } from "stream/consumers";
 import MonitoriaPrismaRepository from "../../repositories/Prisma/MonitoriaPrismaRepository.js";
 import { Monitoria } from "@prisma/client";
 import { eventEmmiter } from "../../utils/events/Evento.js";
+import { UsuarioLogado } from "../../models/UsuarioLogado.js";
 
 interface MonitoriaInput {
   nome_monitoria: string;
@@ -294,14 +295,31 @@ class MonitoriaService {
     return dadosMonitoria;
   } // atualiza monitoria
 
-  async delete(id: string): Promise<Monitoria> {
-    const monitoriaDados = await this._monitoriaRepository.delete(id);
+  async delete(id: string, user : UsuarioLogado): Promise<Monitoria> {
 
-    if (!monitoriaDados) {
-      throw new Error("MONITORIA_INEXISTENTE");
+    const monitoriaExistente = await this._monitoriaRepository.getById(id)
+
+    if (!monitoriaExistente){
+      throw new Error ("MONITORIA_INEXISTENTE")
     }
 
-    return monitoriaDados;
+    if (monitoriaExistente.monitorId != user.id && user.perfil != "ADMIN"){
+      throw new Error ("EXCLUSAO_NAO_PERMITIDA")
+    }
+
+    const monitoriaExcluida = await this._monitoriaRepository.delete(id);
+
+    eventEmmiter.emit("Monitoria:Deletada", {
+      usuarioId: user.id,
+      monitoriaId: monitoriaExcluida.id,
+      monitoria: {
+        nome: monitoriaExcluida.nome_monitoria,
+        data: monitoriaExcluida.inicio.toISOString(),
+        monitorId: monitoriaExcluida.monitorId,
+      },
+    });
+
+    return monitoriaExcluida;
   } // deleta monitoria
 }
